data-gathering/requirements.txt
PyGithub
pymongo
pymongo[srv]
alive_progress